rendered show of quiz attempts

+++ backend/database/migrations/2023_03_08_100104_create_quiz_attempts_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
     {
         Schema::create('quiz_attempts', function (Blueprint $table) {
             $table->id();
-            $table->integer('score');
+            $table->decimal('score',5,4);
             $table->foreignId('quiz_id')->constrained()->cascadeOnDelete();
             $table->foreignId('user_id')->constrained()->cascadeOnDelete();
             $table->timestamps();
